lambda/apigateway/requirements.txt
minio
pinecone
python-dotenv
google-genai
pdfplumber
docx2txt
